h2vFiles/requirements.txt
matplotlib==3.1.1
networkx==2.4
numpy==1.16.5
pandas==0.23.4
scikit-image==0.15.0
scikit-learn==0.23.1
scipy==1.4.1
tqdm==4.36.1
pytorch-nlp==0.5.0
pyyaml==5.4.1
pyverilog
pyunpack
patool
pathlib
pydot